Demonstrate Fix
The bug was the way `_super.included` was called. ember-cli breaks if you don't bind `this` here. Yeah, I know, that seems bad and I don't know why it does that.

I also removed the `publicAssetURL` option to autoImport, which wasn't doing anything and doesn't really belong in an addon anyway. It's purpose is to say where assets are hosted in production (if you happen to move your Javascript files to weird URLs).